server: isolate rTorrent specific functions
