+++ server/services/rTorrent/constants/methodCallConfigs/index.ts
