+++ server/services/rTorrent/constants/methodCallConfigs/torrentPeer.ts
