+++ server/services/rTorrent/constants/methodCallConfigs/torrentTracker.ts
