+++ server/services/rTorrent/constants/methodCallConfigs/transferSummary.ts
