+++ server/services/rTorrent/util/fileTreeUtil.js
