+++ server/services/rTorrent/util/scgiUtil.js
