+++ server/services/torrentService.ts
@@ -4,8 +4,7 @@ import type {TorrentProperties, TorrentListDiff, TorrentListSummary} from '@shar
 
 import BaseService from './BaseService';
 import config from '../../config';
-
-import {hasTorrentFinished} from '../util/torrentPropertiesUtil';
+import hasTorrentFinished from '../util/torrentPropertiesUtil';
 
 interface TorrentServiceEvents {
   FETCH_TORRENT_LIST_SUCCESS: () => void;
